% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{calc_bandwidth}
\alias{calc_bandwidth}
\title{Calculate bandwidth for the ggridges plot}
\usage{
calc_bandwidth(data, target, group)
}
\description{
Calculate bandwidth for the ggridges plot
}
\details{
Stolen from {ggridges} packages
}
