% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{cdf_exposure}
\alias{cdf_exposure}
\title{Plot of Cummulative exposure distribution}
\usage{
cdf_exposure(data, var_exp, var_group = NULL, ref_value)
}
\arguments{
\item{data}{The tbl_exposure() for Individuals}

\item{var_exp}{String. The variable name for the exposure values. This depends on the scenario}

\item{var_group}{String.Is there a grouping in the data? Usually demographic}

\item{ref_value}{Numeric. The reference value if any}
}
\description{
Plot of Cummulative exposure distribution
}
\details{
Plots either single or multigroup CDF depending if you
supply a `var_group`
}
