% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_tables.R
\name{create_tbl_subjects}
\alias{create_tbl_subjects}
\title{The table of subjecs}
\usage{
create_tbl_subjects(consumption, ...)
}
\arguments{
\item{consumption}{The uploaded consumption file}

\item{...}{Unused yet}
}
\description{
The table of subjecs
}
