% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{pdf_exposure}
\alias{pdf_exposure}
\title{Histogram for the exposure statistics}
\usage{
pdf_exposure(data, var_exp, bins, accuracy = 0.1, digits = 3)
}
\arguments{
\item{data}{The Exposure table by subject}

\item{var_exp}{Variaable that holds the individuaal exposure}

\item{accuracy}{Rounding digits in percentages}

\item{digits}{Round digits in exposure values in x - axis}

\item{binsHow}{How many bins?}
}
\description{
Histogram for the exposure statistics
}
