% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{pdf_exposureDemo}
\alias{pdf_exposureDemo}
\title{PDF exposure by Group
 Compare distribtuion across groups
 @param data The data
 @param var_exp  String. The exposure variable.Differes depending the scenario
 @param var_group String. The grouping variable
 @param bandwidth Numeric.The bandwidth for the density plots
 @param scale Numeric
 @param ref_value Numeric. The reference value
 @noRd}
\usage{
pdf_exposureDemo(data, var_exp, var_group, scale = 1.1, ref_value)
}
\description{
PDF exposure by Group
 Compare distribtuion across groups
 @param data The data
 @param var_exp  String. The exposure variable.Differes depending the scenario
 @param var_group String. The grouping variable
 @param bandwidth Numeric.The bandwidth for the density plots
 @param scale Numeric
 @param ref_value Numeric. The reference value
 @noRd
}
