% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{pdf_exposureDemo2}
\alias{pdf_exposureDemo2}
\title{Pdf ecosure by group - Freq_poly}
\usage{
pdf_exposureDemo2(data, var_exp, var_group, bins, digits = 3)
}
\description{
Pdf ecosure by group - Freq_poly
}
