% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{pdf_exposureDemo}
\alias{pdf_exposureDemo}
\title{PDF exposure by Group
 Compare distribtuion across groups}
\usage{
pdf_exposureDemo(data, var_exp, var_group, bandwith, scale = 1.1, ref_value)
}
\description{
PDF exposure by Group
 Compare distribtuion across groups
}
